Faltan cositas

+++ Unidad1/Ejercicio07TratamientoCSV/CarreraFinal.java
@@ -4,7 +4,7 @@ public class CarreraFinal extends TipoCarrera {
     private boolean extraPoint;
     private String fastestLap;
 
-    public CarreraFinal(String track, int position, int no, String driver, String team, int startingGrid, int laps, String time, float points, boolean extraPoint, String fastestLap) {
+    public CarreraFinal(String track, int position, int no, String driver, String team, int startingGrid, int laps, String time, double points, boolean extraPoint, String fastestLap) {
         super(track, position, no, driver, team, startingGrid, laps, time, points);
         this.extraPoint = extraPoint;
         this.fastestLap = fastestLap;
