+++ Unidad1/Ejercicio07TratamientoCSV/SprintCarrera.java
@@ -2,7 +2,7 @@ package Ejercicio07TratamientoCSV;
 
 public class SprintCarrera extends TipoCarrera {
 
-    public SprintCarrera(String track, int position, int no, String driver, String team, int startingGrid, int laps, String time, float points) {
+    public SprintCarrera(String track, int position, int no, String driver, String team, int startingGrid, int laps, String time, double points) {
         super(track, position, no, driver, team, startingGrid, laps, time, points);
     }
 }
